-adicionado opção no instalador para criar menu(-m --newmenu) -adicionado arquivo xprofile -adicionado numlockx em xprofile -adicionado xcompmgr em xprofile -removido picom em openbox autostart -adicionado arquivos de configuração do cool-retro-term

+++ teste/testar.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 
-rm /home/work/projetos/dotfiles.tar
-tar -cf /home/work/projetos/dotfiles.tar /home/work/projetos/dotfiles
+if [ -e "/home/work/projetos/dotfiles.tar" ];then
+	rm /home/work/projetos/dotfiles.tar
+fi
+cd /home/work/projetos
+#tar -cf /home/work/projetos/dotfiles.tar /home/work/projetos/dotfiles
+tar -cf dotfiles.tar dotfiles
 echo "removendo usuario..."
 sudo userdel -r tt
 echo "adicioando usuario"
@@ -9,7 +13,7 @@ sudo useradd -m tt
 echo "senha para novo usuario "
 sudo passwd tt
 #sudo echo "git clone https://github.com/dnc0/dotfiles" > /home/tt/script.sh
-sudo cp /home/work/projetos/dotfiles.tar /home/tt
-sudo tar -xf /home/tt/dotfiles.tar -C /home/tt
-echo "login do usuario"
-su tt
+eval "sudo tar -xf dotfiles.tar -C /home/tt"
+#echo "login do usuario"
+#su tt
+echo "ok"
